ft (database) Ajout de l'update de serveur de table
Création des méthodes d'update de table,
ainsi que les méthodes permettant d'affecter
un serveur à une table

+++ src/main/java/fr/ul/miage/groupe7/projetrestaurant/Database/TableDAO.java
@@ -2,9 +2,17 @@ package fr.ul.miage.groupe7.projetrestaurant.Database;
 
 import com.mongodb.client.result.DeleteResult;
 import org.bson.Document;
+import org.bson.conversions.Bson;
 import org.bson.types.ObjectId;
 
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
+
 import static com.mongodb.client.model.Filters.eq;
+import static com.mongodb.client.model.Updates.set;
 
 public class TableDAO extends DAO<Table> {
 
@@ -25,13 +33,27 @@ public class TableDAO extends DAO<Table> {
                 : new Table(d);
     }
 
+    public List<Table> findByServeurNull(){
+        var d = connect.find(eq("serveur",null))
+                .into(new ArrayList<>());
+        return (d.isEmpty()) ? Collections.emptyList()
+                : d.stream().map(Table::new).collect(Collectors.toList());
+    }
+
+    public List<Table> findAll(){
+        var d = connect.find()
+                .into(new ArrayList<>());
+        return (d.isEmpty()) ? Collections.emptyList()
+                : d.stream().map(Table::new).collect(Collectors.toList());
+    }
+
     @Override
     public Table create(Table obj) {
         Document d;
         d = new Document("etage", obj.getEtage())
                 .append("numero", obj.getNumero())
                 .append("etat", obj.getEtat().name())
-                .append("serveur", obj.getServeur().get_id());
+                .append("serveur", (obj.getServeur() != null) ? obj.getServeur().get_id() : null);
 
         var insert =  connect.insertOne(d);
         ObjectId id = insert.getInsertedId().asObjectId().getValue();
@@ -41,17 +63,33 @@ public class TableDAO extends DAO<Table> {
 
     @Override
     public Table update(Table obj) {
-        return null;
+        if(obj.get_id() != null) {
+            List<Bson> updates = new ArrayList<>();
+            //UPDATE du serveur de la table
+            Utilisateurs serveur = this.find(obj.get_id()).getServeur();
+            if ( !(obj.getServeur() != null && obj.getServeur().get_id() == null) && !Objects.equals(serveur, obj.getServeur())) {
+                updates.add((obj.getServeur() == null) ? set("serveur", null)
+                        : set("serveur", obj.getServeur().get_id()));
+            }
+            int etage = this.find(obj.get_id()).getEtage();
+            if ( obj.getEtage() != etage) {
+                updates.add(set("etage", etage));
+            }
+            if (!(updates.isEmpty()))
+                connect.updateOne(eq(obj.get_id()), updates);
+            return find(obj.get_id());
+        }
+        System.err.println("La table que vous cherchez à mettre à jour n'existe pas");
+        return obj;
     }
 
     @Override
     public boolean delete(Table obj) {
-        DeleteResult res = connect.deleteOne(eq("_id", obj.get_id()));
+        if(obj != null) {
+            DeleteResult res = connect.deleteOne(eq("_id", obj.get_id()));
 
-        if(res.getDeletedCount() == 0){
-            return false;
-        }else{
-            return true;
+            return res.getDeletedCount() != 0;
         }
+        return false;
     }
 }
